refactor table,body variables
