+++ src/variables/body/table.tsx
@@ -11,9 +11,7 @@ import React, { useState } from 'react';
 
 import { Variables } from '../index';
 
-import useTbody from './useTbody';
-
-export class VariablesTable extends ReactWidget {
+export class Table extends ReactWidget {
   constructor(model: Variables.IModel) {
     super();
     this.model = model;
@@ -83,4 +81,40 @@ const TableComponent = ({ model }: { model: Variables.IModel }) => {
       <TableBody />
     </div>
   );
+};
+
+const useTbody = (items: Array<any>, defaultState: any, id: string = '') => {
+  const [state, setState] = useState(defaultState);
+
+  const setClassIcon = (typeOf: string) => {
+    return typeOf === 'class' ? 'jp-ClassIcon' : 'jp-VariableIcon';
+  };
+
+  const List = () => (
+    <div style={{ overflow: 'auto' }}>
+      <table>
+        <tbody>
+          {items.map(item => (
+            <tr
+              key={item.name}
+              onClick={e => setState(item)}
+              className={id + (state === item ? ' selected' : '')}
+            >
+              <td style={{ paddingLeft: `${12}px`, width: `${25}%` }}>
+                <span
+                  className={`jp-Icon jp-Icon-16 ${setClassIcon(item.value)}`}
+                ></span>
+                {item.name}
+              </td>
+              <td style={{ paddingLeft: `${12}px`, width: `${75}%` }}>
+                {item.value}
+              </td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+
+  return [state, List, setState];
 };
