--- src/variables/body/useOutsideClick.tsx
@@ -1,19 +1,0 @@
-import { useEffect } from 'react';
-
-const useOutsideClick = (ref: any, callback: any) => {
-  const handleClickOutside = (e: Event) => {
-    if (ref.current && !ref.current.contains(e.target)) {
-      callback();
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      //unbind
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  });
-};
-
-export default useOutsideClick;
